Dirty extraction of posts to own crate
