+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["server", "video-toolkit", "tracing-common"]
+members = ["posts", "server", "video-toolkit", "tracing-common"]
 
 [workspace.dependencies]
 clap = { version = "4.2", features = ["derive"] }
@@ -33,3 +33,8 @@ serde = "1.0"
 serde_json = "1.0"
 
 chrono = { version = "0.4.23", default-features = false }
+
+rss = { version = "2.0", features = ["validation"] }
+markdown = "1.0.0-alpha"
+include_dir = { version = "0.7.3", features = ["metadata", "glob"] }
+path-absolutize = "3.0"
