+++ posts/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "posts"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rss = { workspace = true }
+miette = { workspace = true }
+chrono = { workspace = true }
+
+serde = { workspace = true }
+serde_yaml = "0.9"
+
+markdown = { workspace = true }
+include_dir = { workspace = true }
+
+path-absolutize = { workspace = true }
