+++ server/src/http_server/mod.rs
@@ -5,25 +5,33 @@ use axum::{
     routing::*,
     Router, Server,
 };
+use chrono::{DateTime, NaiveTime, Utc};
 use include_dir::*;
 use miette::{Context, IntoDiagnostic};
+use posts::{
+    blog::{BlogPost, ToCanonicalPath},
+    date::PostedOn,
+    past_streams::PastStream,
+    til::TilPost,
+    title::Title,
+    Post,
+};
 use std::{net::SocketAddr, sync::Arc};
 use tower_http::trace::TraceLayer;
 
-use crate::{
-    posts::blog::{BlogPosts, ToCanonicalPath},
-    AppState,
-};
+use crate::{AppConfig, AppState};
 pub use config::*;
 use errors::*;
+
+use self::{pages::blog::md::IntoHtml, templates::ShortDesc};
 
 pub(crate) mod cmd;
 
 pub(crate) mod pages {
     pub mod blog;
     pub mod home;
-    pub mod til;
     pub mod streams;
+    pub mod til;
 }
 
 mod api {
@@ -69,4 +77,55 @@ pub(crate) async fn run_axum(config: AppState) -> miette::Result<()> {
         .await
         .into_diagnostic()
         .wrap_err("Failed to run server")
+}
+
+pub(crate) trait LinkTo {
+    fn relative_link(&self) -> String;
+
+    fn absolute_link(&self, config: &AppConfig) -> String {
+        config.app_url(&self.relative_link())
+    }
+}
+
+impl LinkTo for BlogPost {
+    fn relative_link(&self) -> String {
+        format!("/posts/{}", self.path.canonical_path())
+    }
+}
+
+impl LinkTo for TilPost {
+    fn relative_link(&self) -> String {
+        format!("/til/{}", self.frontmatter.slug)
+    }
+}
+
+impl LinkTo for PastStream {
+    fn relative_link(&self) -> String {
+        format!("/streams/{}", self.frontmatter.date)
+    }
+}
+
+pub(crate) trait ToRssItem {
+    fn to_rss_item(&self, state: &AppState) -> rss::Item;
+}
+
+impl<FrontMatter> ToRssItem for Post<FrontMatter>
+where
+    FrontMatter: PostedOn + Title,
+    Post<FrontMatter>: LinkTo,
+{
+    fn to_rss_item(&self, state: &AppState) -> rss::Item {
+        let link = self.absolute_link(&state.app);
+
+        let posted_on: DateTime<Utc> = self.posted_on().and_time(NaiveTime::MIN).and_utc();
+        let formatted_date = posted_on.to_rfc2822();
+
+        rss::ItemBuilder::default()
+            .title(Some(self.title().to_string()))
+            .link(Some(link))
+            .description(self.short_description())
+            .pub_date(Some(formatted_date))
+            .content(Some(self.markdown().ast.0.into_html(state).into_string()))
+            .build()
+    }
 }
