+++ server/src/http_server/pages/blog/md/plain.rs
@@ -70,7 +70,8 @@ impl IntoPlainText for Root {
 
 #[cfg(test)]
 mod test {
-    use crate::posts::MarkdownAst;
+
+    use posts::MarkdownAst;
 
     use super::*;
